Refactor to organise some functions
API calls relating to a user's drinks cabinet have been relocated to a
new file, as has the standardResponse function, and functions have been
ordered alphabetically to make them easier to find.
