+++ routes/SendResponse.js
@@ -1,0 +1,10 @@
+class SendResponse {
+  static standardResponse(response, error, result) {
+    if (error) {
+      response.send(error);
+    }
+    response.json(result);
+  }
+}
+
+module.exports = SendResponse;
